requirements.txt
coloredlogs~=15.0.1
aiogram~=3.1.1
motor~=3.3.1
python-dotenv~=1.0.0